requirements.txt
anthropic>=0.30.0
pytrends>=4.9.2
google-api-python-client>=2.130.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=3.0.0
schedule>=1.2.0
gunicorn
line-bot-sdk
Pillow>=10.0.0
